Update Empty Activity Android Application project spec, fix testing
