+++ unified-prototype/unified-plugin/plugin-android/src/main/resources/templates/android-application-basic-activity/app/build.gradle.dcl
@@ -1,0 +1,34 @@
+androidApplication {
+    namespace = "org.example.app"
+
+    viewBinding {
+        enabled = true
+    }
+
+    dependencies {
+        implementation("androidx.core:core-ktx:1.10.1")
+        implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.6.1")
+
+        implementation("androidx.appcompat:appcompat:1.6.1")
+        implementation("com.google.android.material:material:1.10.0")
+
+        implementation("androidx.constraintlayout:constraintlayout:2.1.4")
+        implementation("androidx.navigation:navigation-fragment-ktx:2.6.0")
+        implementation("androidx.navigation:navigation-ui-ktx:2.6.0")
+    }
+
+    buildTypes {
+        release {
+            minify {
+                enabled = false
+            }
+
+            defaultProguardFile {
+                name = "proguard-android-optimize.txt"
+            }
+            proguardFile {
+                name = "proguard-rules.pro"
+            }
+        }
+    }
+}
