+++ unified-prototype/unified-plugin/plugin-android/src/main/resources/templates/android-application-basic-activity/app/src/androidTest/java/org/example/app/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.tomtresansky.myapplication
+package org.example.app
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.tomtresansky.myapplication", appContext.packageName)
+        assertEquals("org.example.app", appContext.packageName)
     }
 }
