+++ unified-prototype/unified-plugin/plugin-android/src/main/resources/templates/android-application-basic-activity/app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.tomtresansky.myapplication.FirstFragment"
+        android:name="org.example.app.FirstFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.tomtresansky.myapplication.SecondFragment"
+        android:name="org.example.app.SecondFragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second">
