+++ unified-prototype/unified-plugin/plugin-android/src/main/resources/templates/android-application-basic-activity/app/src/test/java/org/example/app/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.tomtresansky.myapplication
+package org.example.app
 
 import org.junit.Test
