--- unified-prototype/unified-plugin/plugin-android/src/main/resources/templates/android-application-basic-activity/build.gradle.kts
@@ -1,5 +1,0 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-plugins {
-    alias(libs.plugins.android.application) apply false
-    alias(libs.plugins.kotlin.android) apply false
-}
