+++ unified-prototype/unified-plugin/plugin-common/src/testFixtures/resources/templates/java-library/build.gradle.dcl
