+++ unified-prototype/unified-plugin/plugin-common/src/testFixtures/resources/templates/java-library/src/main/java/com/example/lib/Library.java
